vrclient: Fix incorrect sized int type for VRInputValueHandle_t.
Fixes a regression in Racket: NX from 9b0930f925009c4dbd30ecd83281cd0c7b91eb5b.

CW-Bug-Id: #22729
CW-Bug-Id: #23147

+++ vrclient_x64/vrinput_manual.c
@@ -4,7 +4,7 @@ WINE_DEFAULT_DEBUG_CHANNEL(vrclient);
 
 #ifndef __x86_64__
 static uint32_t ivrinput_get_digital_action_data( uint64_t action_handle, void *action_data, uint32_t action_data_size,
-                                                  uint32_t restrict_to_device, unsigned int version )
+                                                  uint64_t restrict_to_device, unsigned int version )
 {
     /* Digital action state change fixup hack. */
     w_InputDigitalActionData_t *data = action_data;
@@ -67,7 +67,7 @@ static uint32_t ivrinput_get_digital_action_data( uint64_t action_handle, void *
 
 uint32_t __thiscall winIVRInput_IVRInput_004_GetDigitalActionData( struct w_steam_iface *_this, uint64_t action,
                                                                    w_InputDigitalActionData_t *pActionData,
-                                                                   uint32_t unActionDataSize, uint32_t ulRestrictToDevice )
+                                                                   uint32_t unActionDataSize, uint64_t ulRestrictToDevice )
 {
     w_InputDigitalActionData_t w_pActionData;
     struct IVRInput_IVRInput_004_GetDigitalActionData_params params =
@@ -95,7 +95,7 @@ uint32_t __thiscall winIVRInput_IVRInput_004_GetDigitalActionData( struct w_stea
 
 uint32_t __thiscall winIVRInput_IVRInput_005_GetDigitalActionData( struct w_steam_iface *_this, uint64_t action,
                                                                    w_InputDigitalActionData_t *pActionData,
-                                                                   uint32_t unActionDataSize, uint32_t ulRestrictToDevice )
+                                                                   uint32_t unActionDataSize, uint64_t ulRestrictToDevice )
 {
     w_InputDigitalActionData_t w_pActionData;
     struct IVRInput_IVRInput_005_GetDigitalActionData_params params =
@@ -123,7 +123,7 @@ uint32_t __thiscall winIVRInput_IVRInput_005_GetDigitalActionData( struct w_stea
 
 uint32_t __thiscall winIVRInput_IVRInput_006_GetDigitalActionData( struct w_steam_iface *_this, uint64_t action,
                                                                    w_InputDigitalActionData_t *pActionData,
-                                                                   uint32_t unActionDataSize, uint32_t ulRestrictToDevice )
+                                                                   uint32_t unActionDataSize, uint64_t ulRestrictToDevice )
 {
     w_InputDigitalActionData_t w_pActionData;
     struct IVRInput_IVRInput_006_GetDigitalActionData_params params =
@@ -151,7 +151,7 @@ uint32_t __thiscall winIVRInput_IVRInput_006_GetDigitalActionData( struct w_stea
 
 uint32_t __thiscall winIVRInput_IVRInput_007_GetDigitalActionData( struct w_steam_iface *_this, uint64_t action,
                                                                    w_InputDigitalActionData_t *pActionData,
-                                                                   uint32_t unActionDataSize, uint32_t ulRestrictToDevice )
+                                                                   uint32_t unActionDataSize, uint64_t ulRestrictToDevice )
 {
     w_InputDigitalActionData_t w_pActionData;
     struct IVRInput_IVRInput_007_GetDigitalActionData_params params =
@@ -179,7 +179,7 @@ uint32_t __thiscall winIVRInput_IVRInput_007_GetDigitalActionData( struct w_stea
 
 uint32_t __thiscall winIVRInput_IVRInput_010_GetDigitalActionData( struct w_steam_iface *_this, uint64_t action,
                                                                    w_InputDigitalActionData_t *pActionData,
-                                                                   uint32_t unActionDataSize, uint32_t ulRestrictToDevice )
+                                                                   uint32_t unActionDataSize, uint64_t ulRestrictToDevice )
 {
     w_InputDigitalActionData_t w_pActionData;
     struct IVRInput_IVRInput_010_GetDigitalActionData_params params =
